+++ features/support/env.rb
@@ -1,4 +1,0 @@
-require 'cucumber/rails'
-
-ActionController::Base.allow_rescue = false
-Cucumber::Rails::Database.javascript_strategy = :truncation