+++ features/support/factory_girl.rb
@@ -1,2 +1,0 @@
-# Enable short factory_girl syntax
-World(FactoryGirl::Syntax::Methods)