+++ features/support/webmock.rb
@@ -1,6 +1,0 @@
-require 'webmock/cucumber'
-WebMock.disable_net_connect!(:allow_localhost => true)
-
-Before do
-  stub_request(:any, 'https://www.google.com/jsapi')
-end