+++ spec/features/dataset_add_document_spec.rb
@@ -1,0 +1,26 @@
+require 'spec_helper'
+
+RSpec.feature 'Adding a dataset to a search', type: :feature do
+  scenario 'when adding a single article' do
+    sign_in_with
+    create_dataset
+
+    visit datasets_path
+    click_link 'Manage'
+
+    expect(page).to have_content('Number of documents: 427')
+
+    visit search_path
+    first(:button, 'More').click
+    click_link 'Add this document to an existing dataset'
+
+    within('.modal-dialog') do
+      click_button 'Add'
+    end
+
+    visit datasets_path
+    click_link 'Manage'
+
+    expect(page).to have_content('Number of documents: 428')
+  end
+end